update replace component
