+++ src/components/replace/Replace.vue
@@ -1,3 +1,4 @@
+<template>
 <transition name="modal">
 <div class="modal-mask" @click="close">
   <div style="height:30%"></div>
@@ -27,3 +28,32 @@
   </div>
 </div>
 </transition>
+</template>
+
+<script>
+module.exports = {
+    data: function() {
+        return {
+            applyToAll:false
+        }
+    },
+    props: ['replace_message', 'replace_body', 'consumer_cancel_func', 'consumer_func', 'showApplyAll'],
+    created: function() {
+    },
+    methods: {
+        close: function() {
+            this.$emit("hide-replace");
+        },
+        no: function() {
+            this.close();
+            this.consumer_cancel_func(this.applyToAll);
+        },
+        yes: function() {
+            this.close();
+            this.consumer_func(this.applyToAll);
+        }
+    }
+}
+</script>
+<style>
+</style>
