+++ src/peergos.js
@@ -1,7 +1,6 @@
 var CodeEditor      = require('./components/code-editor');
 var Password        = require('./components/password');
 var PasswordUtil    = require('./components/passwordutil');
-var Replace         = require('./components/replace');
 var SelectCreate    = require('./components/select-create');
 var Spinner         = require('./components/spinner');
 var Text            = require('./components/viewers/text-viewer');
@@ -15,7 +14,6 @@ var App  			= require('./components/App.vue');
 // Loading components
 Vue.component('code-editor', Vue.extend(CodeEditor));
 Vue.component('password', Vue.extend(PasswordUtil).extend(Password));
-Vue.component('replace', Vue.extend(Replace));
 Vue.component('select-create', Vue.extend(SelectCreate));
 Vue.component('spinner', Vue.extend(Spinner));
 Vue.component('text-viewer', Vue.extend(Text));
